Enable user attribute mapping
Users in keycloak have their user details which are a fixed number of
fields, and they also have attributes, which can be added to in a
key-value format. The former (user details) could be added in the WSFED
client mappers to be transmitted in the SAML tokens, but the latter
could not.

The class which would allow these attributes to be added already
existed (SAMLUserAttributeStatementMapper), but was not accessible from
the webfront. This correction enables the class to be accessed by
recording it as a ProtocolMapper. This is simply done by adding the full
classpath of the class in the appropriate file (the file modified with
this commit).